// Wohnungsgeberbestätigung/src/Wohnungsgeberbestaetigung/src/app/dto/wohnungsgeberbestaetigung-dto.ts
export class WohnungsgeberbestaetigungDto {
  nameBeauftragt: string;
  wohnortBeauftragt: string;
  nameEigentuemer: string;
  wohnortEigentuemer: string;
  nameWohnungsgeber: string;
  wohnortWohnungsgeber: string;
  adresse: string;
  einzugsbestaetigung: string;
  tagDesEinzugs: string;
  email: string;
}